Renamed "poem lines" to "puzzles"
The name "poem line" was confusing because there are also lines of the poem (text UI) that show at the end via the PoemViewer class. Earlier, I was going to show the lines of the poem after the clue was solved (in addition to altogether at the end), so the naming made more sense initially. "Puzzle" is a better name because these classes involve the UI and validation for solving the clues.