Fixed bug where last tile couldn't get deleted
Because the indexing of currentTile was maxed out at length of the word minus 1 (to prevent out of range issues), the furthest tile that could be cleared was the second to last tile (since deleting requires going back a tile). By letting current tile be the length of the word (but preventing incrementing beyond that), the last tile of the word can be deleted smoothly.